requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
PyPDF2
unstructured
openai
python-dotenv
pydantic
